==> data_src/notebooks/01_extract_data.ipynb <==
# ivofernandes thesis 

#install dependencies
import sys

!{sys.executable} -m pip install wget

#create path for landing_zone
import os
import wget


try:
    #this is folder is ignored in .gitignore
    os.makedirs('../landing_zone')
    print('landing_zone created')
except FileExistsError:
    print('folder already exists\n')
    try:
        print('deleting leftovers...\n')
        os.remove('../landing_zone/foods.csv')
    except:
        print('something went wrong, please verify locally')
            
try:
    print('Pulling data..')
    !wget -O ../landing_zone/foods.csv 'https://static.openfoodfacts.org/data/en.openfoodfacts.org.products.csv'
except:
    print('error')
